restructuring
HtmlStyleKeeper, MdStyleKeeper and styleInfo moved to ./Styles

TagSearchers renamed to Seachers
ITagSearcher and implementing classes moved to Searchers
EscapeSearcher moved to Searchers

Removed unused constructor in Tag.cs
