+++ cs/MarkdownTask/MarkdownTask/Searchers/HeaderTagSearcher.cs
@@ -1,7 +1,8 @@
 ﻿using System.Collections.Generic;
+using MarkdownTask.Styles;
 using MarkdownTask.Tags;
 
-namespace MarkdownTask.TagSearchers
+namespace MarkdownTask.Searchers
 {
     public class HeaderTagSearcher : ITagSearcher
     {
