+++ cs/MarkdownTask/MarkdownTaskTests/SearchersTests/EscapeSearcherTests.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using FluentAssertions;
-using MarkdownTask;
+using MarkdownTask.Searchers;
 using NUnit.Framework;
 
 namespace MarkdownTaskTests.SearchersTests
