+++ cs/MarkdownTask/MarkdownTaskTests/SearchersTests/SearchersCompositionTests.cs
@@ -2,9 +2,9 @@
 using System.Collections.Generic;
 using System.Linq;
 using FluentAssertions;
-using MarkdownTask;
+using MarkdownTask.Searchers;
+using MarkdownTask.Styles;
 using MarkdownTask.Tags;
-using MarkdownTask.TagSearchers;
 using NUnit.Framework;
 
 namespace MarkdownTaskTests.SearchersTests
